Removed drw stuff, cursors and schemes.
Also, I'm tired of having a different config.h and config.def.h, so
the latter was removed and Makefile updated accordingly.

I also removed any remnants of schemes and cursors, switching to just
the variables {un,}focusbordercolor. The drw files were completely
removed, and the only leftover was one color function for the borders.

I also removed the Client mon field. This enabled us to remove a
couple of functions, and I was also able to remove the motionnotify
stuff.